Confirm account deletion and minor refactor
